Refactor valuetype-related functions
